refactor: ratelimit パッケージを責務別に分割しファイル名のtypoを修正
- platform/ratelimit を platform/httpratelimit にリネーム（HTTP ミドルウェア用 Redis スライディングウィンドウ）
- shared/ratelimiter を platform/clientratelimit にリネーム（外部 API クライアント用 in-memory）
- 空ディレクトリ clientratelimit/ httpratelimit/ に実体を投入し、shared/ ディレクトリを削除
- twelvedata/dto/time_series_reponse.go を time_series_response.go に修正
- 依存する 8 ファイルの import と識別子を更新
- .golangci.yml depguard の deny エントリを httpratelimit に更新
- CLAUDE.md / README.md / feature README のディレクトリ構造記述と Mermaid 図を更新
