+++ internal/feature/candles/adapters/twelvedata/dto/time_series_response.go
