+++ internal/platform/httpratelimit/middleware.go
@@ -1,4 +1,4 @@
-package ratelimit
+package httpratelimit
 
 import (
 	"fmt"
