+++ internal/platform/httpratelimit/middleware_test.go
@@ -1,4 +1,4 @@
-package ratelimit
+package httpratelimit
 
 import (
 	"encoding/json"
