+++ internal/platform/httpratelimit/ratelimit_test.go
@@ -1,4 +1,4 @@
-package ratelimit
+package httpratelimit
 
 import (
 	"context"
